Move query limit enforcement logic to its own file
In preparation for making it more complicated, factor the logic for enforcing query limits in the server out to its own class.
